fixed up some more bugs
loading related
