+++ UltimateRecyclerView/app/src/main/java/com/marshalchen/ultimaterecyclerview/demo/loadmoredemo/BasicFunctions.java
@@ -1,0 +1,184 @@
+package com.marshalchen.ultimaterecyclerview.demo.loadmoredemo;
+
+import android.graphics.drawable.Drawable;
+import android.os.Bundle;
+import android.os.Handler;
+import android.support.v4.widget.SwipeRefreshLayout;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+import android.support.v7.widget.Toolbar;
+import android.view.ActionMode;
+import android.view.View;
+
+import com.marshalchen.ultimaterecyclerview.DragDropTouchListener;
+import com.marshalchen.ultimaterecyclerview.ItemTouchListenerAdapter;
+import com.marshalchen.ultimaterecyclerview.ObservableScrollState;
+import com.marshalchen.ultimaterecyclerview.ObservableScrollViewCallbacks;
+import com.marshalchen.ultimaterecyclerview.URLogs;
+import com.marshalchen.ultimaterecyclerview.UltimateRecyclerView;
+import com.marshalchen.ultimaterecyclerview.demo.R;
+import com.marshalchen.ultimaterecyclerview.uiUtils.ScrollSmoothLineaerLayoutManager;
+
+/**
+ * Created by hesk on 19/2/16.
+ */
+public abstract class BasicFunctions extends AppCompatActivity {
+
+    protected UltimateRecyclerView ultimateRecyclerView;
+
+    protected void enableParallaxHeader() {
+        ultimateRecyclerView.setParallaxHeader(getLayoutInflater().inflate(R.layout.parallax_recyclerview_header, ultimateRecyclerView.mRecyclerView, false));
+        ultimateRecyclerView.setOnParallaxScroll(new UltimateRecyclerView.OnParallaxScroll() {
+            @Override
+            public void onParallaxScroll(float percentage, float offset, View parallax) {
+                Drawable c = toolbar.getBackground();
+                c.setAlpha(Math.round(127 + percentage * 128));
+                toolbar.setBackgroundDrawable(c);
+            }
+        });
+    }
+
+    protected void enableLoadMore() {
+        // StickyRecyclerHeadersDecoration headersDecor = new StickyRecyclerHeadersDecoration(simpleRecyclerViewAdapter);
+        // ultimateRecyclerView.addItemDecoration(headersDecor);
+        ultimateRecyclerView.setLoadMoreView(R.layout.custom_bottom_progressbar);
+
+        ultimateRecyclerView.setOnLoadMoreListener(new UltimateRecyclerView.OnLoadMoreListener() {
+            @Override
+            public void loadMore(int itemsCount, final int maxLastVisiblePosition) {
+                status_progress = true;
+                Handler handler = new Handler();
+                handler.postDelayed(new Runnable() {
+                    public void run() {
+
+                        onLoadmore();
+                        status_progress = false;
+                    }
+                }, 500);
+            }
+        });
+
+    }
+
+    protected abstract void onLoadmore();
+
+    protected abstract void onRefresh();
+
+    protected void enableRefresh() {
+        ultimateRecyclerView.setDefaultOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                new Handler().postDelayed(new Runnable() {
+                    @Override
+                    public void run() {
+                        onRefresh();
+                    }
+                }, 1000);
+            }
+        });
+        //        ultimateRecyclerView.setDefaultSwipeToRefreshColorScheme(getResources().getColor(android.R.color.holo_blue_bright),
+//                getResources().getColor(android.R.color.holo_green_light),
+//                getResources().getColor(android.R.color.holo_orange_light),
+//                getResources().getColor(android.R.color.holo_red_light));
+
+    }
+
+    protected LinearLayoutManager setupLinearLayoutMgr() {
+        linearLayoutManager = new ScrollSmoothLineaerLayoutManager(this, LinearLayoutManager.VERTICAL, false, 300);
+        return linearLayoutManager;
+    }
+
+
+    protected void enableScrollControl() {
+        ultimateRecyclerView.setScrollViewCallbacks(new ObservableScrollViewCallbacks() {
+            @Override
+            public void onScrollChanged(int scrollY, boolean firstScroll, boolean dragging) {
+                URLogs.d("onScrollChanged: " + dragging);
+            }
+
+            @Override
+            public void onDownMotionEvent() {
+
+            }
+
+            @Override
+            public void onUpOrCancelMotionEvent(ObservableScrollState observableScrollState) {
+                URLogs.d("onUpOrCancelMotionEvent");
+                if (observableScrollState == ObservableScrollState.UP) {
+                    ultimateRecyclerView.hideToolbar(toolbar, ultimateRecyclerView, getScreenHeight());
+                    ultimateRecyclerView.hideFloatingActionMenu();
+                } else if (observableScrollState == ObservableScrollState.DOWN) {
+                    ultimateRecyclerView.showToolbar(toolbar, ultimateRecyclerView, getScreenHeight());
+                    ultimateRecyclerView.showFloatingActionMenu();
+                }
+            }
+        });
+
+        ultimateRecyclerView.showFloatingButtonView();
+    }
+
+    protected void enableEmptyViewPolicy() {
+        ultimateRecyclerView.setEmptyView(R.layout.empty_view, UltimateRecyclerView.EMPTY_KEEP_HEADER_AND_LOARMORE);
+        //  ultimateRecyclerView.setEmptyView(R.layout.empty_view, UltimateRecyclerView.EMPTY_KEEP_HEADER);
+        // ultimateRecyclerView.setEmptyView(R.layout.empty_view, UltimateRecyclerView.EMPTY_SHOW_LOADMORE_ONLY);
+        // ultimateRecyclerView.setEmptyView(R.layout.empty_view, UltimateRecyclerView.EMPTY_CLEAR_ALL);
+    }
+
+    protected void enableSwipe() {
+
+    }
+
+    protected void enableItemClick() {
+        ItemTouchListenerAdapter itemTouchListenerAdapter = new ItemTouchListenerAdapter(ultimateRecyclerView.mRecyclerView,
+                new ItemTouchListenerAdapter.RecyclerViewOnItemClickListener() {
+                    @Override
+                    public void onItemClick(RecyclerView parent, View clickedView, int position) {
+                    }
+
+                    @Override
+                    public void onItemLongClick(RecyclerView parent, View clickedView, int position) {
+                        URLogs.d("onItemLongClick()" + isDrag);
+                        if (isDrag) {
+                            URLogs.d("onItemLongClick()" + isDrag);
+                            //   dragDropTouchListener.startDrag();
+                            //   ultimateRecyclerView.enableDefaultSwipeRefresh(false);
+                        }
+
+                    }
+                });
+        ultimateRecyclerView.mRecyclerView.addOnItemTouchListener(itemTouchListenerAdapter);
+    }
+
+    protected void getURV() {
+
+    }
+    protected ActionMode actionMode;
+    protected Toolbar toolbar;
+    protected LinearLayoutManager linearLayoutManager;
+    private int moreNum = 2;
+
+
+    boolean isDrag = true, isEnableAutoLoadMore = true, status_progress = false;
+    private DragDropTouchListener dragDropTouchListener;
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_loadmore);
+        toolbar = (Toolbar) findViewById(R.id.tool_bar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
+        ultimateRecyclerView = (UltimateRecyclerView) findViewById(R.id.ultimate_recycler_view);
+        doURV(ultimateRecyclerView);
+        //setupSpinnerSelection();
+    }
+
+    public int getScreenHeight() {
+        return findViewById(android.R.id.content).getHeight();
+    }
+
+
+    protected abstract void doURV(UltimateRecyclerView urv);
+}
